README.md: Added a rough description of the changes over the versions.
src/manifest.json:
Version bump.

src/background.js:
Fixed a bug introduced in v0.3 where we add an empty array to the end of the list and add our "singles" to the largest set of tabs.  This change corrects that by adding the empty array to the start of the list, and our "singles" go in there. (This bugfix did not entirely resolve our bunching issue with domains with two tabs).
Added logic that alters the method of adding tabs to the final array depending on the size of the bucket.  This brings back some of the problems with v0.1, but it isn't so bad.  We suspect this version may be reasonably optimal because in the left 10% or so of tabs the very high count and very low count of domains are over represented (rather than a strictly homogeneous weave).