This commit doesn't function.
This commit is done while in the process of refactoring AND testing yet another set of weave routines.

This commit is to keep some history around should we wish to revisit some code we are about to delete.

In addition to refactoring, we are adding a couple 'RoundRobin' weaves, which won't result in a homogeneous list of tabs, but may prove useful.
We are also adding "10%" weaves, where we consider the left-most 10%+sqrt(tabCount) of tabs.  These are quite useful for testing, but may also be useful for our own usage patterns.